Parse MakeBucket error and try multiple regions for play.minio.io
Attempt to connect to play.minio.io on either us-east-1 or us-east-2. Fail if the client can't create a bucket for either region.

Error will be parsed for the MakeBucket command in the TestS3Storage test. Current behavior is to ignore all errors. New behavior is to ignore nil error or already-exists error, try a new region if it returns wrong-region error, and fail for all other errors.