Add fio install and robustness-tool-tests makefile target
Fio install target is added, a robustness tools target invokes it and sets the FIO_EXE environment variable, launches tests from the robustness tools directory (for now).

Currently fio is only installed if the `apt-get` binary is found, otherwise install skips to avoid error. The `FIO_EXE` env is populated with the `fio` executable if it is found, empty otherwise. The tests will skip if that env is empty.