manifest: refactored management of committed manifests into separate component
This is very similar to content.CommittedReadManager and allows
multiple writes to coexist and share the underlying committed manifest
manager.

Changed compaction to only include previously-committed manifests and
not any pending ones.

Added locking assertions for extra safety.

This intentionally does as little code move as possible to facilitate
reviewing.