cli: added 'snapshot pin' command
This allows pinning and unpinning of snapshots, which will protect
them from deletion. Any number of pins can be put on a manifest
to indicate reasons why the snapshot is being pinned.

This is similar to retention labels, but is manually managed.

Usage:

```
$ kopia snapshot create path --pin pin1 --pin pin2
$ kopia snapshot pin <manifest-id> --add pin3
$ kopia snapshot pin <manifest-id> --remove pin2
```

Instead of `<manifest-id>` you can also specify directory object ID.

Fixes #1518