ui: add support for authentication using one-time tokens passed via URL
This only works for http://localhost addresses and allows opening of
Kopia UI without the user having to log in.

The server maintains a set of short-lived, single-use tokens,
which are checked on each request. If a request arrives which has
`uiAuthToken` set to a valid token, it gets exchanged for a JWT session
cookie, which will authorize the UI user for the duration of the
session. Session cookie is short-lived and auto-extends on each request.

The server will invalidate the token immediately to prevent reuse.

One-time tokens can be generated by the new API: `POST /api/v1/ui-token`
and are valid for maximum of one minute, just enough to open web browser
to open the UI.

To enable this, pass `--ui-auth-token-ttl=30s` or similar duration.

The intended (FUTURE) flows are:

1. Kopia UI shell (this is similar to what KopiaUI does today but in a
   future without Electron we won't be able to auto-login and work with
   self-signed TLS, so we need to switch to http://localhost URIs)

   - ui shell launches kopia server on random port passing
     `--random-password` parameter and captures the random password as
    it's printed to stderr
   - it generates token by POST http://localhost:port/api/v1/ui-token
     authenticated as `kopia` with the password captured before
   - opens http://localhost:port/?uiAuthToken=<token> in user's browser

2. In the future we can add new flags that will start a one-off server
   and automatically open browser for certain CLI operations to avoid
   having to pass tons of flags in the CLI or to better visualize:

   `kopia repository connect --ui`
   `kopia snapshot list <source> --ui`
   `kopia policy edit <source> --ui`