feat(repository): implement shared async writer pool for object manager
This gets rid of managing asynchronous upload parallelism from the
uploader on a per-file basis.

The effect is slight speed up of snapshots (around 10%), but the
trade-off is slight increase in memory usage (because more memory is
needed for in-flight buffers). Memory usage be improved in #1963 for
overall reduced usage compared to 0.10.7.