docs(site): update instructions on creating issues (#2053)
New issues should all be created in the main (`kopia/kopia`) project,
as that project is the most watched. Keeping all issues in one
central location will help us track and address them more
efficiently. We can use issue labels to indicate that an issue
relates to the UI or other auxiliary project.