docs(site): Cleaning up Release Notes
I think it will be more efficient if we just direct people to the GitHub releases page for anyone that wants to read release notes. I don't see the need to maintain the same information in two different places, plus this section of the site has not been updated for a while.