Add files via upload
We do not have to use this background image. But if we do, we should find some place to credit https://unsplash.com/photos/oyXis2kALVg. The image is free for us to use and we don't need to credit per their license -- but it is the right thing to do.