feat(cli): ensure `--checkpoint-interval` is <=45min
Also hide the flag, since it's not recommended to be tweaked
anyway.

The value of <=45m is very important for safety of the garbage
collection algorithms - too long an interval between checkpoints could
mean that GC treats contents in the middle of being uploaded
as unused, because they are not reachable from any snapshots or
checkpoints.

Fixes #2193