fix(providers): disabled rclone write-back behavior
There are many solutions floating online intended to help speed up
`rclone` performance with high-latency backends, such as Google Drive or
OneDrive.

One of the often-suggested ones is `--vfs-cache-mode`, which creates
an on-disk cache and performs uploads in the background
(asynchronous write-back).

This is not necessarily wrong when a single client is used, but may
cause problems when multiple Kopia clients share repository and
some of the writes are reordered or never happen, in particular in
case of crashes.

Fixes #2216